Fixed TestWebService and TestDaoGeneticAlteration
Added disableForeignKeyCheck for all truncate operations on tables that
have foreign keys.
Escaped the column name VALUES in DaoGenericAlteration because it’s a
reserved word in SQL so MySQL was throwing error.
Removed deleteAllRecords from TestDaoCancerStudy because it was
deleting the seed data as well and some other tests use that seed data,
so instead of deleting all records, I just delete the records added in
the same test case.